now it's possible to handle wishlist items even for unregistered users, using session

+++ includes/class-alg-wc-wish-list-item.php
@@ -21,8 +21,14 @@ if (!class_exists('Alg_WC_Wish_List_Item')) {
 		 * @param type $item_id
 		 * @return int|false Meta ID on success, false on failure.
 		 */
-		public static function add_item_to_wish_list($user_id, $item_id) {
-			$response = add_user_meta($user_id, Alg_WC_Wish_List_User_Metas::WISH_LIST_ITEM, $item_id, false);
+		public static function add_item_to_wish_list($item_id, $user_id = null) {
+			if ($user_id) {
+				$response = add_user_meta($user_id, Alg_WC_Wish_List_User_Metas::WISH_LIST_ITEM, $item_id, false);
+			} else {
+				$_SESSION[Alg_WC_Wish_List_Session_Vars::WISH_LIST] = isset($_SESSION[Alg_WC_Wish_List_Session_Vars::WISH_LIST]) ? $_SESSION[Alg_WC_Wish_List_Session_Vars::WISH_LIST] : array();
+				array_push($_SESSION[Alg_WC_Wish_List_Session_Vars::WISH_LIST], $item_id);
+				$response = $item_id;
+			}
 			return $response;
 		}
 
@@ -33,8 +39,15 @@ if (!class_exists('Alg_WC_Wish_List_Item')) {
 		 * @param type $item_id
 		 * @return bool True on success, false on failure.
 		 */
-		public static function remove_item_from_wish_list($user_id, $item_id) {
-			$response = delete_user_meta($user_id, Alg_WC_Wish_List_User_Metas::WISH_LIST_ITEM, $item_id, false);
+		public static function remove_item_from_wish_list($item_id, $user_id = null) {
+			if ($user_id) {
+				$response = delete_user_meta($user_id, Alg_WC_Wish_List_User_Metas::WISH_LIST_ITEM, $item_id, false);
+			} else {
+				$_SESSION[Alg_WC_Wish_List_Session_Vars::WISH_LIST] = isset($_SESSION[Alg_WC_Wish_List_Session_Vars::WISH_LIST]) ? $_SESSION[Alg_WC_Wish_List_Session_Vars::WISH_LIST] : array();
+				$index = array_search($item_id, $_SESSION[Alg_WC_Wish_List_Session_Vars::WISH_LIST]);
+				unset($_SESSION[Alg_WC_Wish_List_Session_Vars::WISH_LIST][$index]);
+				$response=true;
+			}
 			return $response;
 		}
 
@@ -44,9 +57,9 @@ if (!class_exists('Alg_WC_Wish_List_Item')) {
 		 * @param type $item_id
 		 * @return boolean
 		 */
-		public static function is_item_in_wish_list($user_id, $item_id) {
-			$wishlisted_items	 = get_user_meta($user_id, Alg_WC_Wish_List_User_Metas::WISH_LIST_ITEM, false);
-			$response			 = false;
+		public static function is_item_in_wish_list($item_id, $user_id = null) {
+			$wishlisted_items = Alg_WC_Wish_List::get_wish_list($user_id);
+			$response = false;
 			if (is_array($wishlisted_items) && count($wishlisted_items) > 0) {
 				$index = array_search($item_id, $wishlisted_items);
 				if ($index === false) {
@@ -63,12 +76,11 @@ if (!class_exists('Alg_WC_Wish_List_Item')) {
 		/**
 		 * Remove or add an Item from User Wishlist
 		 */
-		public static function toggle_item_from_wish_list($user_id, $item_id) {
-			$wishlisted_items = get_user_meta($user_id, Alg_WC_Wish_List_User_Metas::WISH_LIST_ITEM, false);
-			if (self::is_item_in_wish_list($user_id, $item_id)) {
-				$response = self::remove_item_from_wish_list($user_id, $item_id);
+		public static function toggle_item_from_wish_list($item_id, $user_id = null) {
+			if (self::is_item_in_wish_list($item_id, $user_id)) {
+				$response = self::remove_item_from_wish_list($item_id, $user_id);
 			} else {
-				$response = self::add_item_to_wish_list($user_id, $item_id);
+				$response = self::add_item_to_wish_list($item_id, $user_id);
 			}
 			return $response;
 		}
